+++ pkg/lang/codegen.go
@@ -218,6 +218,8 @@ func stringConst(e Expr) (string, bool) {
 					return capitalize(v), true
 				case "title":
 					return title(v), true
+				case "swapcase":
+					return swapcase(v), true
 				case "lower":
 					return strings.ToLower(v), true
 				case "strip":
@@ -462,6 +464,8 @@ func (g *irGen) stringVal(e Expr) (string, bool) {
 			return "", false
 		}
 		switch attr.Name.Value {
+		case "swapcase":
+		return swapcase(v), true
 		case "title":
 		return title(v), true
 		case "capitalize":
@@ -1274,6 +1278,8 @@ func (g *irGen) call(b *strings.Builder, c *Call) (string, error) {
 				v = capitalize(v)
 			case "title":
 				v = title(v)
+			case "swapcase":
+				v = swapcase(v)
 		case "lower":
 			v = strings.ToLower(v)
 		case "strip":